+++ app/src/main/kotlin/com/looker/kenko/ui/planEdit/components/DaySwitcher.kt
@@ -104,10 +104,11 @@ fun DaySwitcher(
                     },
                 ) { day ->
                     Text(
-                        text = kenkoDayName(dayOfWeek = day),
+                        text = dayName(dayOfWeek = day),
                         style = MaterialTheme.typography.labelLarge,
                     )
                 }
+
             },
         )
         Spacer(modifier = Modifier.width(6.dp))